+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## Unreleased
 - Initial Release after splitting api and models
+- add transformer for classifying the state of boxes